backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==0.11.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
temporalio==1.5.1
assemblyai==0.21.0
pytube==15.0.0
yt-dlp
stripe==7.8.0
discord.py==2.3.2
asyncio-mqtt==0.16.1
httpx==0.25.2
python-dotenv==1.0.0
asyncpg==0.29.0
google-generativeai==0.3.2
